Remoção do rabbit entre outras coisas ;) (pretendo implementar um circuit breaker)

+++ src/main/java/com/pieropan/app/dto/PaymentProcessor.java
@@ -1,4 +1,6 @@
 package com.pieropan.app.dto;
 
-public record PaymentProcessor(int totalRequests, Double totalAmount) {
+import java.math.BigDecimal;
+
+public record PaymentProcessor(int totalRequests, BigDecimal totalAmount) {
 }
