+++ src/main/java/com/pieropan/app/dto/PaymentProcessorRequest.java
@@ -1,6 +1,7 @@
 package com.pieropan.app.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
-public record PaymentProcessorRequest(String correlationId, Double amount, Instant requestedAt) {
+public record PaymentProcessorRequest(String correlationId, BigDecimal amount, Instant requestedAt) {
 }
